WIP code for the Rust union runtime
I'm leaving it at that as the proper implementation would not only be
complex but also require multistep postprocessing from the convertor. It
is much easier for now hoist all the literals and mark all unions as
untagged. Later I can get back to it and make it right and enable behind
an option (or not).